# Increase the maximum size of glyph atlas