--- a/js/data/bucket/symbol_bucket.js
+++ b/js/data/bucket/symbol_bucket.js
@@ -60,12 +60,12 @@ var layoutVertexArrayType = new Bucket.VertexArrayType([{
     components: 2,
     type: 'Int16'
 }, {
-    name: 'a_data1',
-    components: 4,
-    type: 'Uint8'
-}, {
-    name: 'a_data2',
+    name: 'a_texture_pos',
     components: 2,
+    type: 'Uint16'
+}, {
+    name: 'a_data',
+    components: 4,
     type: 'Uint8'
 }]);
 
@@ -73,20 +73,23 @@ var elementArrayType = new Bucket.ElementArrayType();
 
 function addVertex(array, x, y, ox, oy, tx, ty, minzoom, maxzoom, labelminzoom, labelangle) {
     return array.emplaceBack(
-            // pos
+            // a_pos
             x,
             y,
-            // offset
-            Math.round(ox * 64), // use 1/64 pixels for placement
+
+            // a_offset
+            Math.round(ox * 64),
             Math.round(oy * 64),
-            // data1
-            tx / 4,                   // tex
-            ty / 4,                   // tex
+
+            // a_texture_pos
+            tx / 4, // x coordinate of symbol on glyph atlas texture
+            ty / 4, // y coordinate of symbol on glyph atlas texture
+
+            // a_data
             (labelminzoom || 0) * 10, // labelminzoom
-            labelangle,               // labelangle
-            // data2
-            (minzoom || 0) * 10,               // minzoom
-            Math.min(maxzoom || 25, 25) * 10); // minzoom
+            labelangle, // labelangle
+            (minzoom || 0) * 10, // minzoom
+            Math.min(maxzoom || 25, 25) * 10); // maxzoom
 }
 
 SymbolBucket.prototype.addCollisionBoxVertex = function(layoutVertexArray, point, extrude, maxZoom, placementZoom) {