--- a/js/symbol/glyph_source.js
+++ b/js/symbol/glyph_source.js
@@ -28,7 +28,7 @@ GlyphSource.prototype.getSimpleGlyphs = function(fontstack, glyphIDs, uid, callb
         this.stacks[fontstack] = {};
     }
     if (this.atlases[fontstack] === undefined) {
-        this.atlases[fontstack] = new GlyphAtlas(128, 128);
+        this.atlases[fontstack] = new GlyphAtlas();
     }
 
     var glyphs = {};