Enlarging the glyph sheet allows CJK characters to render and not overflow. Works with https://github.com/mapbox/mapbox-gl-test-suite/pull/131 tests.

Performance of this PR on bench:
Buffer, 803 ms
FPS, 60 frames per second

cc @amyleew