@bhousel 2016-08-02
Yay!  Been waiting a long time for this..

1. I think this might necessitate a corresponding change in https://github.com/mapbox/mapbox-gl-shaders
2. If you're going to do this, I'd raise the limit to 4096x4096.