@lucaswoj 2016-08-02
> 1. I think this might necessitate a corresponding change in https://github.com/mapbox/mapbox-gl-shaders

Fortunately, shaders are not coupled to the bit-width of these attributes. https://github.com/mapbox/mapbox-gl-shaders/blob/master/src/icon.vertex.glsl#L11-L12