@lucaswoj 2016-08-02
> Performance of this PR on bench:
> Buffer, 803 ms
> FPS, 60 frames per second

These benchmarks are machine-specific and therefore relative numbers. Can you re-run these benchmarks against both this branch and `master`? (This'll be easier when we get around to https://github.com/mapbox/mapbox-gl-js/issues/2405)