@bhousel 2016-08-03
> Fortunately, shaders are not coupled to the bit-width of these attributes. https://github.com/mapbox/mapbox-gl-shaders/blob/master/src/icon.vertex.glsl#L11-L12

Isn't the glyph shader
https://github.com/mapbox/mapbox-gl-shaders/blob/master/src/sdf.vertex.glsl

But either way, that is helpful that the bit width doesn't need to change in the shader code.

> The size of this array has a huge effect on GL JS performance.
> 
> There are a couple of members of data1 that don't need their size increased and no members of data2 that need their size increased.

If performance is a concern, could we dynamically adjust `layoutVertexArrayType` from `Uint8` to `Uint16` in realtime as the atlas increases in size?  We only need it to be `Uint16` for very large symbol and glyph atlases.