@lucaswoj 2016-08-03
> If performance is a concern, could we dynamically adjust `layoutVertexArrayType` from `Uint8` to `Uint16` in realtime as the atlas increases in size?  We only need it to be `Uint16` for very large symbol and glyph atlases.

This is a great idea and is theoretically possible. Implementing it will require some bigger architectural changes. Let's keep this idea in our back pocket.