@xrwang 2016-08-03
Thanks for all the detailed feedback! Updated this PR per comments. Bench results on master versus branch below. Please let me know if I've missed / misunderstood something : )

**master**
---
**buffer**: 795 ms

**fps**: 60 frames per second

**frame-duration**: 
5.4 ms per frame at zoom 4. 0% of frames took longer than 16ms.
5.2 ms per frame at zoom 8. 0% of frames took longer than 16ms.
7 ms per frame at zoom 11. 0% of frames took longer than 16ms.
8.6 ms per frame at zoom 13. 0% of frames took longer than 16ms.
8.3 ms per frame at zoom 15. 0% of frames took longer than 16ms.
5.6 ms per frame at zoom 17. 0% of frames took longer than 16ms.

avg: 6.7 ms per frame. 0% of frames took longer than 16ms.


**this pr**
---

**buffer**: 833 ms

**fps**: 60 frames per second

**frame-duration**:
5.5 ms per frame at zoom 4. 0% of frames took longer than 16ms.
5.2 ms per frame at zoom 8. 0% of frames took longer than 16ms.
7.1 ms per frame at zoom 11. 0% of frames took longer than 16ms.
9.3 ms per frame at zoom 13. 0% of frames took longer than 16ms.
8.1 ms per frame at zoom 15. 0% of frames took longer than 16ms.
5.2 ms per frame at zoom 17. 0% of frames took longer than 16ms.

avg: 6.7 ms per frame. 0% of frames took longer than 16ms.