@lucaswoj 2016-08-03
This is looking great @xrwang! A 5% slowdown in buffer creation time (795 ms -> 833 ms) is a good starting point. Do you think there's anything we can do to make the slowdown even smaller? Did you verify that the slowdown is related to the increase in `StructArray` size?