@xrwang 2016-08-03
@lucaswoj, yup, I can definitely confirm that the slow down is related to the increase in `StructArray` size. Though from benching several times, it might also be related to the glyph atlas size itself, ie 2048 vs 4096? Although technically that shouldn't affect it, only the `StructArray` size should ? Anyhow, happy to keep it at 2048 for now to veer on the cautious side. benched several times below, and this branch with uint16, 2048 seems very close to the master branch's buffer time. 

**benching sequence, results in order***
1. master, buffer: 774ms

2. this branch, with uint 16, 4096 glyph atlas, buffer: 830ms

3. this branch, with uint 8, 4096 glyph atlas, buffer: 777 ms

4. this branch, with uint16, 2048 glyph atlas, buffer:  766 ms *

5. master, buffer: 765 ms

6. this branch with uint16, 2048 glyph atlas, buffer: 792 ms *

7. master, buffer: 784 ms