@lucaswoj 2016-08-03
This looks great! Thank you for doing these benchmarks @xrwang. 

>  Though from benching several times, it might also be related to the glyph atlas size itself, ie 2048 vs 4096? Although technically that shouldn't affect it, only the `StructArray` size should ?

It is possible and looks to be the case that the glyph atlas size has a large effect on performance too. If a width of 2048 is enough for real world usage with CJK fonts then I'm happy with this result  and almost ready to ship this PR.

# Next Steps

 - [ ] confirm that a glyph atlas size of 2048 is big enough for real-world usage with CJK fonts
 - [ ] look for low-hanging preformance fruit in `GlyphAtlas` (specifically the "realloc" algorithm)
 - [ ] merge https://github.com/mapbox/mapbox-gl-test-suite/pull/131 into `master`
 - [ ] update `mapbox-gl-test-suite` in `package.json` in this PR to include https://github.com/mapbox/mapbox-gl-test-suite/pull/131
 - [ ] merge this PR into master