@xrwang 2016-08-05
cool, updated per comments. The name change for a_texture_pos is over here: https://github.com/mapbox/mapbox-gl-shaders/pull/22

Tests over on https://github.com/mapbox/mapbox-gl-test-suite/pull/131 should be ready.

Benched with the resizing by 4 (instead of 2) and performance improved a lot!

**master**
806 ms

**this branch**
796 ms

**master**
814 ms

**this-branch**
822 ms