@brunoabinader 2016-08-05
Created https://github.com/mapbox/mapbox-gl-native/issues/5873 to track these changes in GL native 👍