@lucaswoj 2016-08-11
# Next Steps

 - [ ] update test-suite SHA in GL JS
 - [ ] rename `a_data2` attribute to `a_data`
 - [ ] update shaders SHA in GL JS
 - [ ] squash & rebase GL JS 
 - [ ] run GL JS benchmarks
 - [ ] do a manual test of GL JS on debug pages
 - [ ] bring GL Native PR in sync with `a_data2` shader changes
 - [ ] 🍏 
 - [ ] 🚢 

# Related PRs

 - https://github.com/mapbox/mapbox-gl-test-suite/pull/131
 - https://github.com/mapbox/mapbox-gl-shaders/pull/22
 - https://github.com/mapbox/mapbox-gl-native/pull/5908/files