@lucaswoj 2016-08-11
Pushing squashed & based version. Previous HEAD was db7c0d8bd85bc1a4e8622e10c902664c248d3f5f