@lucaswoj 2016-08-11
# Branch Benchmarks

**buffer:** 960 ms
**fps:** 60 fps
**frame-duration:** 7 ms, 0% > 16ms
**query-point:** 0.67 ms
**query-box:** 50.55 ms
**geojson-setdata-small:** 11 ms
**geojson-setdata-large:** 114 ms

# `master` Benchmarks

**buffer:** 922 ms
**fps:** 60 fps
**frame-duration:** 6.9 ms, 0% > 16ms
**query-point:** 0.71 ms
**query-box:** 50.73 ms
**geojson-setdata-small:** 11 ms
**geojson-setdata-large:** 111 ms